requirements.txt
fastapi
uvicorn
gunicorn
supabase
cryptography
requests
pillow
python-multipart
pydantic-settings
google-generativeai
